# docs/reference/generated/openapi.json
{
  "components": {
    "schemas": {
      "ApiResponse": {
        "additionalProperties": false,
        "properties": {
          "authority_required": {
            "items": {
              "type": "string"
            },
            "title": "Authority Required",
            "type": "array"
          },
          "claims": {
            "additionalProperties": true,
            "title": "Claims",
            "type": "object"
          },
          "code": {
            "title": "Code",
            "type": "string"
          },
          "effect_class": {
            "title": "Effect Class",
            "type": "string"
          },
          "generation_digest": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Generation Digest"
          },
          "job_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Job Id"
          },
          "next_safe_commands": {
            "items": {
              "items": {
                "type": "string"
              },
              "type": "array"
            },
            "title": "Next Safe Commands",
            "type": "array"
          },
          "objects_written": {
            "items": {
              "type": "string"
            },
            "title": "Objects Written",
            "type": "array"
          },
          "quarantined_objects": {
            "items": {
              "type": "string"
            },
            "title": "Quarantined Objects",
            "type": "array"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "tenant_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tenant Id"
          },
          "trace_id": {
            "title": "Trace Id",
            "type": "string"
          },
          "unknowns": {
            "items": {
              "type": "string"
            },
            "title": "Unknowns",
            "type": "array"
          },
          "workspace_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Workspace Id"
          }
        },
        "required": [
          "status",
          "code",
          "effect_class",
          "trace_id"
        ],
        "title": "ApiResponse",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "WorkspaceCreate": {
        "additionalProperties": false,
        "properties": {
          "genesis_envelope_fingerprint": {
            "pattern": "^sha256:[0-9a-f]{64}$",
            "title": "Genesis Envelope Fingerprint",
            "type": "string"
          },
          "root_spki_fingerprint": {
            "pattern": "^sha256:[0-9a-f]{64}$",
            "title": "Root Spki Fingerprint",
            "type": "string"
          },
          "workspace_id": {
            "pattern": "^[A-Za-z0-9][A-Za-z0-9._:-]{0,127}$",
            "title": "Workspace Id",
            "type": "string"
          }
        },
        "required": [
          "workspace_id",
          "root_spki_fingerprint",
          "genesis_envelope_fingerprint"
        ],
        "title": "WorkspaceCreate",
        "type": "object"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "CPCF Evidence-Control API",
    "version": "0.6.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/jobs/{job_id}": {
      "get": {
        "operationId": "job_status_v1_jobs__job_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Job Status"
      }
    },
...
    "/v1/schemas/{kind}": {
      "get": {
        "operationId": "schema_v1_schemas__kind__get",
        "parameters": [
          {
            "in": "path",
            "name": "kind",
            "required": true,
            "schema": {
              "title": "Kind",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Schema"
      }
    },
    "/v1/workspaces": {
      "post": {
        "operationId": "create_workspace_v1_workspaces_post",
        "parameters": [
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "maxLength": 128,
              "minLength": 16,
              "title": "Idempotency-Key",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WorkspaceCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Workspace"
      }
    },
    "/v1/workspaces/{workspace_id}": {
      "get": {
        "operationId": "workspace_status_v1_workspaces__workspace_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "workspace_id",
            "required": true,
            "schema": {
              "title": "Workspace Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Workspace Status"
      }
    },
    "/v1/workspaces/{workspace_id}/analyses": {
      "post": {
        "operationId": "start_analysis_v1_workspaces__workspace_id__analyses_post",
        "parameters": [
          {
            "in": "path",
            "name": "workspace_id",
            "required": true,
            "schema": {
              "title": "Workspace Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "maxLength": 128,
              "minLength": 16,
              "title": "Idempotency-Key",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "If-Match",
            "required": true,
            "schema": {
              "title": "If-Match",
              "type": "string"
            }
          }
        ],
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Start Analysis"
      }
    },
    "/v1/workspaces/{workspace_id}/cas/sha256/{hex_digest}": {
      "put": {
        "description": "Upload bounded immutable bytes; the object remains quarantined until admission.",
        "operationId": "upload_cas_object_v1_workspaces__workspace_id__cas_sha256__hex_digest__put",
        "parameters": [
          {
            "in": "path",
            "name": "workspace_id",
            "required": true,
            "schema": {
              "title": "Workspace Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "hex_digest",
            "required": true,
            "schema": {
              "title": "Hex Digest",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "maxLength": 128,
              "minLength": 16,
              "title": "Idempotency-Key",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "If-Match",
            "required": true,
            "schema": {
              "title": "If-Match",
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Upload Cas Object"
      }
    },
    "/v1/workspaces/{workspace_id}/onboarding": {
      "get": {
        "operationId": "onboard_v1_workspaces__workspace_id__onboarding_get",
        "parameters": [
          {
            "in": "path",
            "name": "workspace_id",
            "required": true,
            "schema": {
              "title": "Workspace Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Onboard"
      }
    }
  }
}